♻️ A585MV code: decompose pr sync steps
